vms-960: alpha-dec-vms is LLP64 + toolchain blocks libc build (empirical)
Built the overlay against the REAL alpha-dec-vms cross toolchain. Two findings:

1) alpha-dec-vms is LLP64, NOT LP64 (task premise was wrong): int=4, long=4,
   long long=8, pointer=8 (-mpointer-size=64), LE, long double=8 (IEEE binary64),
   wchar_t=unsigned int. `long` is 32-bit by VMS ABI. Overlay adapted to LLP64:
   alltypes decouples _Addr/_Int64/_Reg from 32-bit long; setjmp.h uses 64-bit
   elements; atomic pointer ll/sc uses a 64-bit tie. musl configures + the C
   compile begins cleanly. The preflight gate asserts this model and stopped the
   original LP64 attempt (honest, no fake).

2) The alpha-dec-vms GCC (cc1 14.2 + binutils 2.43 GAS) cannot assemble its own
   output for standard C-RTL function names: it decc$-prefixes them and emits
   `.pdesc decc$strlen..en` (entry label is strlen..en -> mismatch) when DEFINING
   and `.linkage decc$t<callee>` when CALLING -> "redefined symbol cannot be used
   on reloc" / "cannot generate BFD_RELOC_16". 702/1353 musl objects (~52%) fail,
   incl. ALL 4 rung-1 targets (strlen/malloc/memcpy/vsnprintf). A non-CRTL name
   (myfunc) assembles fine. This is a cc1/binutils (GCC-lane) bug that sabotages
   defining/calling libc functions -> no libc builds until fixed. Out of scope
   for the musl overlay; route to conductor 3-way (shared toolchain).

Working asm fixes proven this pass: VMS/EVAX procedure model for setjmp/longjmp
(.ent/.base/.frame 4-arg/.prologue no-arg/.link/.pdesc/.end), _start without
ELF .weak/.hidden, syscall.h.in comment no longer leaks the NR token into the
generated header. build-musl.sh now: LLP64 preflight, tries full libc.a, falls
back to a labeled PARTIAL archive of valid objects, enumerates the failing set,
and exits non-zero (honest red) because the 4 targets are absent.

RUNG 1 STATUS: overlay complete + correct; genuine libc.a BLOCKED on the cc1
decc$/.pdesc toolchain bug. PR to main for the record; not merging.

+++ tools/cross-alpha-vms/musl-arch/arch/alpha-dec-vms/atomic_arch.h
@@ -50,7 +50,9 @@ static inline void *a_ll_p(volatile void *p)
 #define a_sc_p a_sc_p
 static inline int a_sc_p(volatile void *p, void *v)
 {
-	long r;
+	/* `long` is 32-bit on alpha-dec-vms (LLP64); the stq_c source register
+	 * must hold the full 64-bit pointer, so use a 64-bit type for the tie. */
+	long long r;
 	__asm__ __volatile__ (
 		"stq_c %1,%0"
 		: "=m"(*(void *volatile *)p), "=&r"(r)
