+++ tools/cross-alpha-vms/musl-arch/arch/alpha-dec-vms/bits/setjmp.h
@@ -7,5 +7,9 @@
  *   [8]      gp ($29)
  *   [9]      sp ($30)
  *   [10..17] float callee-saved $f2-$f9
+ *
+ * Elements MUST be 64-bit: setjmp.s stores quadwords (stq/stt). `long` is only
+ * 32-bit on alpha-dec-vms (LLP64), so use `long long` here or the buffer is half
+ * the size the asm writes.
  */
-typedef unsigned long __jmp_buf[18];
+typedef unsigned long long __jmp_buf[18];
